// src/domain/types.ts
export interface Order {
  id: string;
  date: string;
  customer: string; // name
  total: number;
  status: string;
  method: string;
  items: number;
}